refactor pages to use new shared components
